--- app/src/components/grid/README.md
@@ -1,14 +1,0 @@
-# Shortcuts
-For mac only. If windows, one may consider changing all `Ctrl` to `Alt`.
-
-## Filtering
-- `Ctrl + F + C`: Remove all filters
-
-## Row Grouping
-- `Ctrl + LeftArrow`: Expand one level.
-- `Ctrl + RightArrow`: Reduce one level.
-
-# Context Menus
-
-## Filtering
-- `Filter equal`: Equivalent to 'column = {value}'.
