+++ app/src/components/grid/StatusBar.tsx
